speedup DRDOS div64 routine
It was noticed that the EDR kernel is noticeably slower on an 8088
than the FreeDOS kernel when performing buffered sub-sector reads via
INT21,3F. The cause was found in the slow div64 routine.

With this patch it is now slightly faster than the FreeDOS kernel
shipped with SvarDOS (2042, TurboC compiled).

https://github.com/SvarDOS/edrdos/issues/116