disk.asm: fix FAT32 detection in INT21,36 handler
This commit in connection with 6e1496ce fixes the bug reported in
http://svardos.org/?p=forum&thread=1736204477

Tracked at https://github.com/SvarDOS/edrdos/issues/128

This bug makes the kernel randomly report wrong free space when the redirector
is involved.  This is a result of choosing a non-deterministic code path
depending on what the dosfat variable contains.

It changes the FAT32 detection in func36 to utilize DDSC_DIRENT instead of the
dosfat variable, which after commit 6e1496ce is initialized to a non-zero
value.