fix directory entry corruption after LFN deletion (first try ;) )
https://github.com/SvarDOS/edrdos/issues/143

When deleting a SFN directory entry after deleting its corresponding
LFN entries, the wrong directory entry was deleted if the last deleted
LFN was in another directory cluster than the SFN, and the SFN entry
was not the first one in a cluster.

The "test bx,bx" instruction inside rd_pcdir then made this function
skip its cluster traversal part and returning the directory entry
of the right position inside the cluster, but of the wrong cluster.

Primarily tests show that the fix works as intended. Further tests for
unintended side-effects should be performed.