in LNAMES and PUBDEF replace leading "D." by '$' dollar sign
NASM doesn't allow to make global a symbol name that begins with
a dollar sign, but it may be needed to link to certain symbols.
The dot is fairly unusual for existing TASM and MASM code, and an
embedded dot doesn't interfere with NASM's local label mechanism.

Multiple "D." may be placed at the beginning of a name, and will
all be replaced. However, any embedded "D." are not replaced.

Reference: https://bugzilla.nasm.us/show_bug.cgi?id=3392922