redir.asm: fix redirector drive `: (for LASTDRIVE=32) imported from https://hg.pushbx.org/ecm/edrdos/rev/8ab6ce75cc58
This requires to run a redirector supporting LASTDRIVE higher than Z:,
eg dosemu2 with patch in https://github.com/dosemu2/dosemu2/pull/2602

FreeCOM also needs a patch to support this. It was merged today, from
the Pull Request in https://github.com/FDOS/freecom/pull/180

lDOS now supports LASTDRIVE=32 as well:
https://hg.pushbx.org/ecm/msdos4/rev/1ed1e77d75b2

MS-DOS v7 needs a similar patch to this changeset as it also changed
the caps bit unconditionally in its DriveFromText function. In that
case it tried to make the drive letter uncapitalised and expected it
to be >= 61h then. Patch in https://pushbx.org/ecm/test/20250926.3/